Feature/Fix
Feature:

Added 4 new stages to the game, each with completed parallax

Added checkpoint feature. When the objective for the current stage is completed, a checkpoint will now be saved to the local device.

Fix:

Moved logic from various scripts and combined them into various managers, this way control is more centralized for specific variables and components